refactor: list 레이아웃 서버/클라이언트 분리 및 metadataBase 추가
